Kill add allBreedsToDB file. Move logic to dbConnection. Schema now can be run whenever and either builds database and adds breeds to database or does nothing
